output through the cli commands
